refactor(checkpoints): name configuration props after their components
